fix: don't use format on user input
since we don't know what format the user will type in the date, we
shouldn't use the format when trying to parse that string.

BREAKING CHANGE: currently the older tests will fail due to the fact
that they are expecting it to parsed with the format. this is making me
think this isn't the right approach.